fix: selected bank appears in the header
