+++ src/pages/Company/Extrajudicial/ExtrajudicialCustomers/CustomersActions/CustomersActions.tsx
@@ -4,8 +4,6 @@ import styled, { css } from 'styled-components'
 import { useLoloContext } from '@/contexts/LoloProvider'
 import Container from '@/ui/Container'
 import TextField from '@/ui/fields/TextField'
-import Select from '@/ui/Select'
-import { SelectItemType } from '@/ui/Select/interfaces'
 import Button from '@/ui/Button/Button'
 import Modal from '@/ui/Modal'
 import useModal from '@/hooks/useModal'
@@ -19,11 +17,10 @@ const CustomersActions = () => {
 
   const {
     client: { customer },
-    bank: { selectedBank, setSelectedBank },
+    bank: { selectedBank },
   } = useLoloContext()
 
   const {
-    clearAllFilters,
     filterSearch: { getSearchFilters, setSearchFilters },
   } = useFiltersContext()
 
@@ -34,24 +31,6 @@ const CustomersActions = () => {
   } = useModal()
 
   const navigate = useNavigate()
-
-  const options: Array<SelectItemType> = customer.customerBanks.map((bank) => {
-    return {
-      key: String(bank.id),
-      label: bank.name,
-    }
-  })
-
-  const onChangeBank = (key: string) => {
-    clearAllFilters()
-
-    const customerBank = customer.customerBanks.find((customerBank) => String(customerBank.id) === key)
-
-    setSelectedBank({
-      idBank: key,
-      idCHB: String(customerBank?.CUSTOMER_HAS_BANK.id),
-    })
-  }
   const searchFilter = getSearchFilters(currentPath)?.opts ?? { filter: '', limit: 50, page: 1 }
 
   const onChangeSearch = (e: React.ChangeEvent<HTMLInputElement>) => {
@@ -64,47 +43,36 @@ const CustomersActions = () => {
   }
 
   return (
-    <StyledContainer width="100%" display="flex" flexDirection="column" alignItems="center" padding="20px" gap="20px">
-      <Container className="actions__select" width="100%">
-        <Select
-          width="100%"
-          label="Seleccionar banco:"
-          value={selectedBank.idBank}
-          options={options}
-          onChange={onChangeBank}
-        />
-      </Container>
-      <Container className="actions__textfield" width="100%" display="flex" alignItems="center" gap="10px">
-        <TextField
-          onChange={onChangeSearch}
-          width="100%"
-          label="Buscar cliente:"
-          placeholder="Buscar cliente por nombre"
-          value={searchFilter.filter}
-          clearInput
-        />
+    <StyledContainer width="100%" display="flex" flexDirection="row" alignItems="center" padding="20px" gap="20px">
+      <TextField
+        onChange={onChangeSearch}
+        width="100%"
+        label="Buscar cliente:"
+        placeholder="Buscar cliente por nombre"
+        value={searchFilter.filter}
+        clearInput
+      />
 
-        <Button
-          className="btn-excel"
-          width="100px"
-          shape="round"
-          trailingIcon="ri-file-excel-line"
-          onClick={showModalManagementExcel}
-          disabled={!selectedBank.idBank}
-          messageTooltip="Exportar excel"
-          permission="P02-01"
-        />
+      <Button
+        className="btn-excel"
+        width="100px"
+        shape="round"
+        trailingIcon="ri-file-excel-line"
+        onClick={showModalManagementExcel}
+        disabled={!selectedBank.idBank}
+        messageTooltip="Exportar excel"
+        permission="P02-01"
+      />
 
-        <Button
-          width="100px"
-          shape="round"
-          trailingIcon="ri-add-fill"
-          onClick={handleClickAddClient}
-          disabled={!selectedBank.idBank}
-          permission="P02-03"
-          messageTooltip="Agregar cliente"
-        />
-      </Container>
+      <Button
+        width="100px"
+        shape="round"
+        trailingIcon="ri-add-fill"
+        onClick={handleClickAddClient}
+        disabled={!selectedBank.idBank}
+        permission="P02-03"
+        messageTooltip="Agregar cliente"
+      />
 
       <Modal
         id="modal-eport-excel"
@@ -128,33 +96,6 @@ const StyledContainer = styled(Container)`
     .btn-excel {
       background-color: ${theme.colors.Success5};
       border: none;
-    }
-    @media ${theme.device.tabletS} {
-      flex-direction: row;
-
-      .actions__textfield .actions_select {
-        width: 50%;
-      }
-    }
-
-    @media ${theme.device.tabletL} {
-      .actions__textfield {
-        width: 60%;
-      }
-
-      .actions__select {
-        width: 40%;
-      }
-    }
-
-    @media ${theme.device.desktopS} {
-      .actions__textfield {
-        width: 70%;
-      }
-
-      .actions__select {
-        width: 30%;
-      }
     }
   `}
 `
